fix nan case for float payload index
