+++ lib/segment/src/index/field_index/numeric_index/numeric_index_key.rs
@@ -1,6 +1,51 @@
-use super::Encodable;
 use crate::index::field_index::histogram::Numericable;
-use crate::types::PointOffsetType;
+use crate::index::key_encoding::{encode_i64_key_ascending, decode_i64_key_ascending, encode_f64_key_ascending, decode_f64_key_ascending};
+use crate::types::{PointOffsetType, IntPayloadType, FloatPayloadType};
+
+pub trait Encodable: Copy {
+    fn encode_key(&self, id: PointOffsetType) -> Vec<u8>;
+
+    fn decode_key(key: &[u8]) -> (PointOffsetType, Self);
+
+    fn cmp_key(&self, other: &Self) -> std::cmp::Ordering;
+}
+
+impl Encodable for IntPayloadType {
+    fn encode_key(&self, id: PointOffsetType) -> Vec<u8> {
+        encode_i64_key_ascending(*self, id)
+    }
+
+    fn decode_key(key: &[u8]) -> (PointOffsetType, Self) {
+        decode_i64_key_ascending(key)
+    }
+
+    fn cmp_key(&self, other: &Self) -> std::cmp::Ordering {
+        self.cmp(other)
+    }
+}
+
+impl Encodable for FloatPayloadType {
+    fn encode_key(&self, id: PointOffsetType) -> Vec<u8> {
+        encode_f64_key_ascending(*self, id)
+    }
+
+    fn decode_key(key: &[u8]) -> (PointOffsetType, Self) {
+        decode_f64_key_ascending(key)
+    }
+
+    fn cmp_key(&self, other: &Self) -> std::cmp::Ordering {
+        if self.is_nan() && other.is_nan() {
+            return std::cmp::Ordering::Equal;
+        }
+        if self.is_nan() {
+            return std::cmp::Ordering::Less;
+        }
+        if other.is_nan() {
+            return std::cmp::Ordering::Greater;
+        }
+        self.partial_cmp(other).unwrap()
+    }
+}
 
 #[derive(Clone, PartialEq)]
 pub struct NumericIndexKey<T: PartialEq + PartialOrd> {
@@ -23,22 +68,21 @@ impl<T: PartialEq + PartialOrd + Encodable + Numericable> NumericIndexKey<T> {
     }
 }
 
-impl<T: PartialEq + PartialOrd> PartialOrd for NumericIndexKey<T> {
+impl<T: PartialEq + PartialOrd + Encodable> PartialOrd for NumericIndexKey<T> {
     fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        match self.key.partial_cmp(&other.key) {
-            Some(core::cmp::Ordering::Equal) => {}
-            None => {}
-            ord => return ord,
+        match self.key.cmp_key(&other.key) {
+            core::cmp::Ordering::Equal => {}
+            ord => return Some(ord),
         }
         Some(self.idx.cmp(&other.idx))
     }
 }
 
-impl<T: PartialEq + PartialOrd> Eq for NumericIndexKey<T> {}
+impl<T: PartialEq + PartialOrd + Encodable> Eq for NumericIndexKey<T> {}
 
-impl<T: PartialEq + PartialOrd> Ord for NumericIndexKey<T> {
+impl<T: PartialEq + PartialOrd + Encodable> Ord for NumericIndexKey<T> {
     fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.partial_cmp(other).unwrap()
+        self.partial_cmp(other).expect("Numeric index key is not comparable")
     }
 }
 
@@ -46,7 +90,6 @@ impl<T: PartialEq + PartialOrd> Ord for NumericIndexKey<T> {
 mod tests {
     use std::collections::BTreeSet;
 
-    use super::super::Encodable;
     use super::*;
     use crate::types::FloatPayloadType;
 
@@ -61,9 +104,9 @@ mod tests {
             NumericIndexKey::new(-0.4, 3),
             NumericIndexKey::new(5.0, 1),
             NumericIndexKey::new(-5.0, 1),
-            //NumericIndexKey::new(f64::NAN, 0),
-            //NumericIndexKey::new(f64::NAN, 2),
-            //NumericIndexKey::new(f64::NAN, 3),
+            NumericIndexKey::new(f64::NAN, 0),
+            NumericIndexKey::new(f64::NAN, 2),
+            NumericIndexKey::new(f64::NAN, 3),
             NumericIndexKey::new(f64::INFINITY, 0),
             NumericIndexKey::new(f64::NEG_INFINITY, 1),
             NumericIndexKey::new(f64::NEG_INFINITY, 2),
@@ -81,9 +124,12 @@ mod tests {
 
         for (b, k) in set_byte.iter().zip(set_keys.iter()) {
             let (decoded_id, decoded_float) = FloatPayloadType::decode_key(b.as_slice());
-            println!("{}:{} \t\t {}:{}", k.key, k.idx, decoded_float, decoded_id);
-            assert_eq!(decoded_id, k.idx);
-            assert_eq!(decoded_float, k.key);
+            if decoded_float.is_nan() && k.key.is_nan() {
+                assert_eq!(decoded_id, k.idx);
+            } else {
+                assert_eq!(decoded_id, k.idx);
+                assert_eq!(decoded_float, k.key);
+            }
         }
     }
 }
